+++ src/index.js
@@ -7,16 +7,11 @@ export default class AddAssetHtmlPlugin {
 
   /* istanbul ignore next: this would be integration tests */
   apply(compiler) {
-    compiler.plugin('compilation', compilation => {
-      compilation.plugin(
-        'html-webpack-plugin-before-html-generation',
-        (htmlPluginData, callback) =>
-          addAllAssetsToCompilation(
-            this.assets,
-            compilation,
-            htmlPluginData,
-            callback,
-          ),
+    compiler.hooks.compilation.tap('AddAssetHtmlPlugin', compilation => {
+      compilation.hooks.htmlWebpackPluginBeforeHtmlGeneration.tapPromise(
+        'AddAssetHtmlPlugin',
+        htmlPluginData =>
+          addAllAssetsToCompilation(this.assets, compilation, htmlPluginData),
       );
     });
   }